Copy changes to overview declaration
We change ‘Complete the declaration’ to “Governing body sign-off and completion”. We realise the application needs to be finalised once the chair has checked it is accurate because a) a contributor could come back and edit answers thereby rendering the signed off invalid. b) we need to send the application out as a PDF for record keeping. c) Previously, an easily-missed ‘submit’ button appeared unexpectadly once the chair completed the application creating an unnecessary step.